Extracts Discpatcher and DIDs
