+++ workspaces/api/src/entry-point/CovidAPI.js
@@ -1,64 +1,5 @@
-import { CovidAPIService, Authenticator } from '../services'
-import { ServiceProxy, StateSerializer } from '@identity-box/utils'
-
-import path from 'path'
-
-class Dispatcher {
-  dids = []
-  stateSerializer
-  authenticator
-
-  constructor ({ serializerFilePath } = {}) {
-    const filePath = serializerFilePath || path.resolve(process.cwd(), 'DIDs.json')
-    this.stateSerializer = new StateSerializer(filePath)
-    this.dids = this.stateSerializer.read() || []
-    this.authenticator = new Authenticator()
-  }
-
-  isRegistred = did => {
-    return this.dids.includes(did)
-  }
-
-  registerDid = (method, did) => {
-    if (!this.isRegistred(did)) {
-      this.dids.push(did)
-      this.stateSerializer.write(this.dids)
-    }
-    return {
-      method: `${method}-response`,
-      params: []
-    }
-  }
-
-  checkDid = (method, did) => {
-    return {
-      method: `${method}-response`,
-      params: [
-        { status: this.isRegistred(did) }
-      ]
-    }
-  }
-
-  dispatch = ({ method, params }) => {
-    console.log('**!!** DISPATCHER **!!**')
-    console.log('received:')
-    console.log('method:', method)
-    console.log('params:', params)
-    switch (method) {
-      case 'authenticate-code1':
-        return this.authenticator.authenticate(params[0])
-      case 'register-did':
-        return this.registerDid(method, params[0].did)
-      case 'check-did':
-        return this.checkDid(method, params[0].did)
-      default:
-        return {
-          method: 'unknown-method',
-          params: []
-        }
-    }
-  }
-}
+import { CovidAPIService, Dispatcher } from '../services'
+import { ServiceProxy } from '@identity-box/utils'
 
 class CovidAPI {
   covidAPIService
