+++ workspaces/api/src/services/index.js
@@ -1,2 +1,4 @@
 export { CovidAPIService } from './CovidAPIService'
 export { Authenticator } from './Authenticator'
+export { Dispatcher } from './Dispatcher'
+export { DIDs } from './DIDs'
